+++ launcher/src/com/genir/aitweaks/launcher/loading/CoreLoader.kt
@@ -42,6 +42,7 @@ class CoreLoader : URLClassLoader(arrayOf(latestCoreURL())) {
             Transformer.newTransform("$core\$BeamWeapon", symbols.beamWeapon.classPath),
             Transformer.newTransform("$core\$ProjectileWeapon", symbols.projectileWeapon.classPath),
             Transformer.newTransform("$core\$LoadingUtils", symbols.loadingUtils.classPath),
+            Transformer.newTransform("$core\$FrontShieldAI", symbols.frontShieldAI.classPath),
 
             // Fields and methods.
             Transformer.newTransform("autofireManager_advance", symbols.autofireManager_advance.name),