+++ src/com/genir/aitweaks/core/state/State.kt
@@ -9,7 +9,7 @@ import com.genir.aitweaks.core.FleetCohesionAI
 import com.genir.aitweaks.core.debug.DebugPlugin
 import com.genir.aitweaks.core.debug.removeGrid
 import com.genir.aitweaks.core.playerassist.AimAssistManager
-import com.genir.aitweaks.core.playerassist.AutoOmniShields
+import com.genir.aitweaks.core.playerassist.ShieldAssist
 import com.genir.aitweaks.core.shipai.AttackCoordinator
 import com.genir.aitweaks.core.shipai.FleetSegmentation
 import com.genir.aitweaks.core.utils.Bounds
@@ -52,7 +52,7 @@ class State : BaseEveryFrameCombatPlugin() {
         fleetCohesion[1],
         accelerationTracker,
         AimAssistManager(),
-        AutoOmniShields(),
+        ShieldAssist(),
         AttackCoordinator(),
     )